+++ cgo/cuvs/brute_force_c.h
@@ -70,10 +70,10 @@ void gpu_brute_force_get_results(gpu_brute_force_search_result_c result_c, uint6
 void gpu_brute_force_free_search_result(gpu_brute_force_search_result_c result_c);
 
 // Returns the capacity of the index buffer
-uint32_t gpu_brute_force_cap(gpu_brute_force_c index_c);
+uint64_t gpu_brute_force_cap(gpu_brute_force_c index_c);
 
 // Returns the current number of vectors in the index
-uint32_t gpu_brute_force_len(gpu_brute_force_c index_c);
+uint64_t gpu_brute_force_len(gpu_brute_force_c index_c);
 
 // Returns info about the index as a JSON string
 char* gpu_brute_force_info(gpu_brute_force_c index_c, void* errmsg);
