+++ cgo/cuvs/ivf_pq_c.cpp
@@ -580,7 +580,7 @@ void gpu_ivf_pq_free_result(gpu_ivf_pq_result_c result_c) {
     delete static_cast<ivf_pq_search_result_t*>(result_c);
 }
 
-uint32_t gpu_ivf_pq_cap(gpu_ivf_pq_c index_c) {
+uint64_t gpu_ivf_pq_cap(gpu_ivf_pq_c index_c) {
     if (!index_c) return 0;
     auto* any = static_cast<gpu_ivf_pq_any_t*>(index_c);
     switch (any->qtype) {
@@ -592,7 +592,7 @@ uint32_t gpu_ivf_pq_cap(gpu_ivf_pq_c index_c) {
     }
 }
 
-uint32_t gpu_ivf_pq_len(gpu_ivf_pq_c index_c) {
+uint64_t gpu_ivf_pq_len(gpu_ivf_pq_c index_c) {
     if (!index_c) return 0;
     auto* any = static_cast<gpu_ivf_pq_any_t*>(index_c);
     switch (any->qtype) {
